Coalesce multiplex_client and ofa_clients
